+++ src/index.ts
@@ -1,11 +1,10 @@
 import { commands, Disposable, ExtensionContext } from "vscode";
 
-import { extensionName } from "./extension-name";
 import { Extension } from "./git/extension";
 import { Logger } from "./util/logger";
 
 const registerCommand = (name: string, callback: () => void): Disposable => {
-    return commands.registerCommand(`${extensionName}.${name}`, callback);
+    return commands.registerCommand(`gitblame.${name}`, callback);
 }
 
 export const activate = (context: ExtensionContext): void => {