+++ src/view.ts
@@ -1,5 +1,4 @@
 import { StatusBarAlignment, StatusBarItem, window } from "vscode";
-import { extensionName } from "./extension-name";
 
 import type { Commit } from "./git/util/stream-parsing";
 
@@ -36,9 +35,19 @@ export class StatusBarView {
         this.out.dispose();
     }
 
+    private command(): string {
+        const action = getProperty("statusBarMessageClickAction");
+
+        if (action === "Open tool URL") {
+            return "gitblame.online"
+        }
+
+        return "gitblame.quickInfo";
+    }
+
     private text(text: string, command: boolean): void {
         this.out.text = "$(git-commit) " + text.trimEnd();
         this.out.tooltip = `git blame${ command ? "" : " - No info about the current line" }`;
-        this.out.command = command ? `${extensionName}.quickInfo` : undefined;
+        this.out.command = command ? this.command() : undefined;
     }
 }